requirements.txt
catboost==1.2.10
category-encoders==2.9.0
fastapi==0.136.0
httpx==0.28.1
ipykernel==7.2.0
lightgbm==4.6.0
loguru==0.7.3
matplotlib==3.10.8
mlxtend==0.24.0
nltk==3.9.4
numpy==2.4.4
pandas==3.0.2
psycopg2-binary==2.9.11
python-dotenv==1.2.2
scikit-learn==1.8.0
seaborn==0.13.2
sqlalchemy==2.0.49
xgboost==3.2.0